+++ src/FoundrySequencerKernel.hpp
@@ -140,17 +140,49 @@ class SeqAttributes {
 //*****************************************************************************
 
 
-class SinglePlayRandom {
-	uint64_t played0, played1;// bit field that says if a given item was already played, max 128 steps or phrases
+class SingleStepRandom {
+	uint32_t played;// bit field that says if a given item was already played, max 32 steps
 	std::vector<uint8_t> candidates;
 	
 	public:
 	
-	void init() {played0 = 0, played1 = 0;}
+	void init() {played = 0x1;}// we always start on (reset to) first step, so mark it as played
 	
 	int getNext(int length) {
 		// get a random unplayed step within length, if none, reset all and choose one randomly
 		int retStep = 0;
+		
+		candidates.clear();
+		for (int i = 0; i < length; i++) {
+			if ( (played & (0x1 << i)) == 0) {
+				candidates.push_back(i);
+			}
+		}
+		if (candidates.empty()) {
+			played = 0;
+			retStep = random::u32() % length;
+		}
+		else {
+			retStep = candidates[random::u32() % candidates.size()];
+		}
+		played |= (0x1 << retStep);
+		
+		return retStep;
+	}
+};// class SingleStepRandom
+
+
+class SinglePhraseRandom {
+	uint64_t played0, played1;// bit field that says if a given item was already played, max 128 phrases
+	std::vector<uint8_t> candidates;
+	
+	public:
+	
+	void init() {played0 = 0x1, played1 = 0;}// we always start on (reset to) first phrase, so mark it as played
+	
+	int getNext(int length) {
+		// get a random unplayed index in tempPhraseIndexes[] within length, if none, reset all and choose one randomly
+		int tpiIndex = 0;
 		
 		candidates.clear();
 		for (int i = 0; i < length; i++) {
@@ -166,24 +198,22 @@ class SinglePlayRandom {
 			}
 		}
 		if (candidates.empty()) {
-			init();
-			retStep = random::u32() % length;
+			played0 = 0, played1 = 0;
+			tpiIndex = random::u32() % length;
 		}
 		else {
-			retStep = candidates[random::u32() % candidates.size()];
+			tpiIndex = candidates[random::u32() % candidates.size()];
 		}
-		if (retStep < 64) {
-			played0 |= (0x1 << retStep);
+		if (tpiIndex < 64) {
+			played0 |= (0x1 << tpiIndex);
 		}
 		else {
-			played1 |= (0x1 << (retStep - 64));
+			played1 |= (0x1 << (tpiIndex - 64));
 		}
 		
-		return retStep;
+		return tpiIndex;
 	}
-
-};// class SinglePlayRandom
-
+};// class SinglePhraseRandom
 
 
 //*****************************************************************************
@@ -235,11 +265,11 @@ class SequencerKernel {
 	unsigned long clockPeriod;// counts number of step() calls upward from last clock (reset after clock processed)
 	int phraseIndexRun;
 	unsigned long phraseIndexRunHistory;
-	SinglePlayRandom singlePhraseRandom;
+	SinglePhraseRandom singlePhraseRandom;
 	bool moveStepIndexRunIgnore;
 	int stepIndexRun;
 	unsigned long stepIndexRunHistory;
-	SinglePlayRandom singleStepRandom;
+	SingleStepRandom singleStepRandom;
 	int ppqnCount;
 	int ppqnLeftToSkip;// used in clock delay
 	int gateCode;// 0 = Low for current pulse of step, 1 = High for current pulse of step, 2 = Clk high pulse, 3 = 1ms trig
@@ -453,7 +483,7 @@ class SequencerKernel {
 	bool movePhraseIndexForeward(bool init, bool rollover);
 	int tempPhraseIndexes[MAX_PHRASES];// used only in next method	
 	void movePhraseIndexRandom(bool init, uint32_t randomValue);	
-	void movePhraseIndexRandomSingle(bool init, uint32_t randomValue);	
+	void movePhraseIndexRandomSingle(bool init);	
 	void movePhraseIndexBrownian(bool init, uint32_t randomValue);	
 	bool movePhraseIndexRun(bool init);
 };// class SequencerKernel